apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
youtube-transcript-api>=1.2.0
pytest>=8.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
textblob>=0.18.0
langdetect>=1.0.9
scikit-learn>=1.0.0
umap-learn>=0.5.0
hdbscan>=0.8.33
sentence-transformers>=2.2.0
nltk>=3.8.0